break project stuff out of index
